+++ holoviews/tests/core/test_decollation.py
@@ -7,7 +7,7 @@ from holoviews.streams import PlotSize, RangeXY, Stream
 from holoviews.testing import assert_element_equal
 from holoviews.tests.test_streams import Sum, Val
 
-from ..utils import ds, ds_skip
+from .._deps import ds, ds_skip
 
 if ds:
     from holoviews.operation.datashader import datashade, spread
