+++ holoviews/tests/operation/test_downsample.py
@@ -7,7 +7,7 @@ import pytest
 import holoviews as hv
 from holoviews.operation.downsample import _ALGORITHMS, downsample1d
 
-from ..utils import tsdownsample_skip
+from .._deps import tsdownsample_skip
 
 algorithms = _ALGORITHMS.copy()
 algorithms.pop("viewport", None)  # viewport return slice(len(data)) no matter the width
