+++ holoviews/tests/plotting/matplotlib/test_overlayplot.py
@@ -4,7 +4,8 @@ import numpy as np
 
 import holoviews as hv
 
-from ...utils import LoggingComparison, mpl, mpl_skip
+from ..._deps import mpl, mpl_skip
+from ...utils import LoggingComparison
 from .test_plot import TestMPLPlot, mpl_renderer
 
 if mpl:
