+++ holoviews/tests/util/test_utils.py
@@ -11,7 +11,8 @@ from holoviews.core.options import OptionTree
 from holoviews.plotting import bokeh
 from holoviews.util import OutputSettings
 
-from ..utils import LoggingComparison, mpl_skip, notebook_skip
+from .._deps import mpl_skip, notebook_skip
+from ..utils import LoggingComparison
 
 BACKENDS = ["matplotlib", "bokeh"]
